Move test_snapshot from test_dashboard.py to modules/snapshot/tests/test_snapshot.py
